fix circular dependency (#28)
* fix circular dependency

* Update CI env vars
