+++ .github/workflows/nodejs_client.yml
@@ -31,7 +31,7 @@ jobs:
     # NPM clean install - Download and build dependencies from scratch
     - run: npm ci
     # NPM run build - Generate a production build of the application (this may have optimizations not included in dev builds)
-    - run: npm run build --if-present
+    - run: CI='' npm run build --if-present
     # Run all unit tests for the back-end
     - run: npm test
