+++ .github/workflows/nodejs_server.yml
@@ -23,6 +23,8 @@ jobs:
       DB_USER: root
       DB_PASSWORD: root
       DATABASE_URL: mysql://root:root@localhost:3306/oncall
+      JWT_SECRET_KEY: canbeanythingtheysay
+      TOKEN_HEADER_KEY: somethingelse
     steps:
     - uses: actions/checkout@v3
     - name: Set up MySQL
