+++ server/.env.example
@@ -12,3 +12,8 @@ DATABASE_URL="mysql://myuser:mypassword@localhost:42333/oncall"
 # You will either have to update user permissions through a SQL call or use "root" user
 
 # The port we are using is 42333 because that is the port mapped to port 3306 (default mysql/mariadb port) on the docker container
+
+# For generating JWT tokens / auth
+JWT_SECRET_KEY = gfg_jwt_secret_key
+  
+TOKEN_HEADER_KEY = gfg_token_header_key
