+++ server/package.json
@@ -4,7 +4,7 @@
   "private": true,
   "scripts": {
     "start": "concurrently \"nodemon ./bin/www\" \"npm run client --prefix ../client\"",
-    "test": "jest --runInBand",
+    "test": "jest --runInBand --detectOpenHandles",
     "client": "npm run client --prefix ../client",
     "server": "nodemon ./bin/www",
     "init": "node ./init/init"
