+++ server/routes/users.js
@@ -4,6 +4,7 @@ const jwt = require('jsonwebtoken');
 const dotenv = require('dotenv');
 const { tokenify } = require('./auth');
 const prisma = require('../prismaClient');
+const { getUserByEmail } = require('../persist/users');
 
 dotenv.config();
 
@@ -13,41 +14,4 @@ router.get('/', async function (req, res, next) {
   res.send(allMyUsers);
 });
 
-
-async function getUserByEmail(email, password) {
-
-  let result = await prisma.user.findFirst({
-    where: {
-      email: email,
-      password: password
-    }
-
-  })
-
-  return result;
-}
-
-
-router.post('/login', async function (req, res, next) {
-  if(!req.body.email || !req.body.password) {
-    res.status(401);
-    res.send("Invalid arguments");
-    return;
-  }
-  const user = await getUserByEmail(req.body.email, req.body.password);
-  if (user == null) {
-
-    res.status(401);
-    res.send("User does not exist!");
-
-  }
-
-  else {
-    const token = await tokenify(user);
-
-    res.send({ token: token });
-  }
-})
-
 module.exports = router;
-module.exports.getUserByEmail=getUserByEmail
