+++ server/routes/users.test.js
@@ -13,7 +13,7 @@ afterAll(() => {
 describe("Test login endpoint", () => {
   test("It should response the GET method", async () => {
     const response = await request(app)
-      .post("/users/login")
+      .post("/auth")
       .send({ email: "cfiande1@unb.ca", password: "mytestpass1" });
     expect(response.statusCode).toBe(200);
     expect(response.headers['content-type']).toBe(
@@ -24,13 +24,13 @@ describe("Test login endpoint", () => {
 
   test("No password", async () => {
     const response = await request(app)
-      .post("/users/login")
+      .post("/auth")
       .send({ email: "cfiande1@unb.ca" })
       expect(response.statusCode).toBe(401);      
   })
   test("Invalid user", async () => {
     const response = await request(app)
-      .post("/users/login")
+      .post("/auth")
       .send({ email: "notarealuser@unb.ca", password: "password" })
       expect(response.statusCode).toBe(401);       
   })
@@ -44,6 +44,6 @@ describe("Test get users endpoint", () => {
       expect(response.headers['content-type']).toBe(
         "application/json; charset=utf-8"
       );
-    //   expect(response.body.token).toBeDefined()
+      expect(response.body.length).toBeGreaterThan(0);
     });
   });
